# swagger.json
{
    "swagger": "2.0","info": 
    {
        "version": "1.0.0",
        "title": "F4Innovation API",
        "description": "Backend API for the F4Innovation platform",
        "license": 
        {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": 
    [
        {
            "name": "Math",
            "description": "Handles math related requests"
        }
    ],
    "schemes": 
    [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": 
    [
        "application/json"
    ],
    "paths" : 
    {
        "/math/show": 
        {
            "get":
            {
                "tags":
                [
                    "Numbers"
                ],
                "description": "Replies with numbers in the system",
                "produces":
                [
                    "application/json"
                ],
                "responses":
                {
                    "200":
                    {
                        "description": "New number was successfully added"
                    }
                }
            }
        },

        "/math/push": 
        {
            "post":
            {
                "tags":
                [
                    "Number push"
                ],
                "description": "Adds a number to the system",
                "parameters": 
                [
                    {
                        "name": "integer",
                        "in": "body",
                        "description": "Integer number to be added to the system",
                        "type": "integer",
                        "required": true
                    }
                ],
                "produces":
                [
                    "application/json"
                ],
                "responses":
                {
                    "200":
                    {
                        "description": "Last number was successfully discarded"
                    }

                }
            }
        },

        "/math/pop": 
        {
            "delete":
            {
                "tags":
                [
                    "Number pop"
                ],
                "description": "Removes last add number from the system",
                "produces":
                [
                    "application/json"
                ],
                "responses":
                {
                    "200":
                    {
                        "description": "New user is logged in"
                    },

                    "400":
                    {
                        "description": "Array is already empty"
                    }
                }
            }
        }
    }
}